Restructuring the project and removing testing code
