+++ 01_network_sysapi/Messenger/MessengerNetworking/include/MessengerNetworking/tcp_core.h
@@ -1,0 +1,17 @@
+#pragma once
+#ifdef _WIN32 
+  #ifndef _WIN32_WINNT
+    #define _WIN32_WINNT 0x0501  /* Windows XP. */
+  #endif
+  #include <winsock2.h>
+  #include <Ws2tcpip.h>
+#else
+  /* Assume that any non-Windows platform uses POSIX-style sockets instead. */
+  #include <sys/socket.h>
+  #include <arpa/inet.h>
+  #include <netdb.h>  /* Needed for getaddrinfo() and freeaddrinfo() */
+  #include <unistd.h> /* Needed for close() */
+#endif
+class tcpCore{
+
+};
