--- 01_network_sysapi/Messenger/MessengerNetworking/include/MessengerNetworking/test.h
@@ -1,4 +1,0 @@
-#pragma once
-enum class tests{
-    one, two, three
-};
