--- 01_network_sysapi/Messenger/MessengerNetworking/src/test.cpp
@@ -1,1 +1,0 @@
-#include "MessengerNetworking/test.h"
